[FLINK-26995][tests] Mark several tests as unit tests
